Update to Asterisk 10.6.0:  this is a bugfix release
The Asterisk Development Team has announced the release of Asterisk 10.6.0.

The release of Asterisk 10.6.0 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* --- format_mp3: Fix a possible crash in mp3_read().

* --- Fix local channel chains optimizing themselves out of a call.

* --- Re-add LastMsgsSent value for SIP peers

* --- Prevent sip_pvt refleak when an ast_channel outlasts its
      corresponding sip_pvt.

* --- Send more accurate identification information in dialog-info SIP
      NOTIFYs.

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-10.6.0

Thank you for your continued support of Asterisk!