net/mldonkey-gui: CONFIGURE_ARGS+= --disable-option-checking
On DragonFly, this package always halted on the configure stage with the
message, "configure: error: unrecognized options: --with-libintl-prefix"

We couldn't find the source of this obsolete configure option, it didn't
come from mldonkey and devel/gettext-lib isn't a dependency.  Adding
--disable-option-checking changed allowed this to be a warning rather than
a fatal error and also allows a successful build.

Background: DragonFly doesn't have libintl in base.