mtools-(at least)4.0.17 tries hard to configure and conditionally set types to use 64bit file offsets where available for seek()ing etc. However, the easy case (sizeof(off_t)>4) is handled incorrectly: mt_size_t is set to size_t - maybe a copy and paste from the fall-back-to-32bit case.
This type is used at least in init.c, when detecting media size and comparing
to the FAT geometry, consequently failing and erroring out with the message
"Big disks not supported on this architecture."

The patch does handle the (e.g. NetBSD) case of 64bit off_t the same as
the case where a off64_t is available (and the other 64bit off_t-equivalent
cases); namely using off_t as mt_size_t.

Thanks to riastradh@ for pointing out where the bug in llong.h was.