Update ruby-kramdown to 0.14.0.
kramdown 0.14.0 released
				Published on Sunday, 16 September 2012

First of all please note that this release contains a backwards-incompatible
change: The syntax for specifying a code language for a code block or code
span has changed. Instead of using lang='CODELANG' one has to use
.language-CODELANG now.

This change has been introduced to avoid problems because the lang="..."
attribute is used by HTML for other purposes than setting the code
language. Furthermore using .language-CODELANG is also proposed by HTML5 and
it seems to be a good way to achieve the needed functionality.

Other changes in this release include the possibility of setting the code
language on the starting line of a fenced code block and a way of excluding
certain headers from the table of contents by assigning the .no_toc class to
them.  Changes

* 2 major changes

  o Code language is now specified via .language-CODELANG instead of
    lang='CODELANG'
  o Implemented support for setting language on fenced code block starting
    line (initial patch by Bran)

* 1 minor change

  o Headers with an ID can be prevented from showing in the TOC by assigning
    the .no_toc class to them (patch by Tim Bates)

* 1 bug fix

  o Numeric instead of symbolic HTML entities are now the default fallback
    (patch by Gioele Barabucci)

kramdown 0.13.8 released
				Published on Friday, 31 August 2012

This release brings two new options (one for adjusting header levels and the
other for enabling/disabling coderay). And the usual bug fixes.  Changes

* 2 minor changes

  o New option header_offset for offsetting all header levels (initial patch
    by Michal Till)
  o New option enable_coderay for enabling/disabling coderay (initial patch by
    Bran)

* 5 bug fixes

  o Reserved HTML characters in abbreviation titles are now correctly output
    (patch by Alex Tomlins)
  o Similar abbreviations (like CSS and CSS3) are now correctly parsed
  o Fixed bug RF#29626: Text of mailto-link was sometimes wrongly obfuscated
    (reported by B Wright)
  o Fixed known Ruby 1.9.3 problem with RakeTest task (patch by Gioele
    Barabucci)
  o Fixed double output of ¡Æmarkdown¡Ç attribute on HTML elements in kramdown
    converter

* 1 documentation change

  o README file is now called README.md and uses kramdown syntax (patch by
    Bran)